test(cli): pin the child process cwd so byte comparisons survive a root run
The byte-for-byte tests spawn a child `bun` process to observe what
`DEFAULT_IO` and the clack wrappers write to the real streams. `Bun.spawn`
inherits the parent's cwd, and the snippet — plus `ui.ts` and `io.ts` behind
it — imports `@clack/prompts` by bare specifier. That package lives in
`apps/cli/node_modules` and is not hoisted, so resolution depends entirely on
where `bun test` was invoked.

The Unit CI job runs with `working-directory: apps/cli` and passed. The
Integration job runs from the repo root, where the child dies with "Cannot
find module '@clack/prompts'" before writing a byte — three failures.

Pin the child's `cwd` to the CLI package root, derived from
`import.meta.dir`.

The guard is what turned this into a red build. Both comparisons assert the
un-injected reference output is non-empty before comparing; without that, two
dead children would have produced `""`, compared equal, and reported green
forever while asserting nothing — in the same job this branch exists to fix.
That is also why the helper is now shared rather than copied into both
suites: a cwd pin that exists twice can regress in one file and go unnoticed,
which is precisely the failure mode above.

Reproduced from the repo root via the preload's `TEST_TIER=0` escape (tier 0
is PGlite, so no Docker): 17 pass / 3 fail before, 20 pass / 0 fail after.

Refs #1180
